requirements.txt
pyautogui==0.9.54
pystray==0.19.5
Pillow==11.1.0
